test(aggregator): test that RPC timeout conditions are handled and proper error msgs are returned.
refactor some test code into separate module for clarity.
