+++ core/Cargo.toml
@@ -18,7 +18,7 @@ hex = { workspace = true, features = ["serde"] }
 hex-literal = { workspace = true }
 lazy_static = { workspace = true, features = ["spin_no_std"] }
 sha2 = { workspace = true }
-risc0-zkvm = { version = "2.0.2", features = [ "prove" ] }
+risc0-zkvm = { version = "2.0.2", features = ["prove"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
 secp256k1 = { workspace = true, features = ["serde", "rand", "std"] }
@@ -70,6 +70,7 @@ citrea-e2e = { workspace = true }
 base64 = { workspace = true }
 bridge-circuit-host = { path = "../bridge-circuit-host" }
 ctor = { workspace = true }
+tokio = { workspace = true, features = ["full", "test-util"] }
 
 [[bin]]
 name = "clementine-cli"
